# dist/swagger.yaml
swagger: '2.0'
info:
  description: >-
    # Introduction

    Welcome to the reference for the Quire REST API!


    <a href="http://en.wikipedia.org/wiki/REST_API" target="_blank">REST</a> is
    a web-service protocol that lends itself to rapid development by using
    everyday HTTP and JSON technology.


    The Quire REST API provides a broad set of operations and resources that:

      * Consistently doing repetitive or tedious tasks.
      * Chaining a process together by responding to changes.
      * Pulling information from other locations like email or Evernote into Quire.
      * Customizing Asana for your team’s processes and workflows.

    Want to share your opinion on how our API works for you? <a
    href="https://quire.io/w/Quire_Feedbacks" target="_blank">Tell us how you
    feel </a>about using our API and what we can do to make it better.


    # API Changelog

    You can find the <a href="https://github.com/quire-api/quire-api"
    target="_blank">Changelog</a> of the API Reference in the Quire Community.


    # Authentication

    ## OAuth v2.0


    Quire recommends that you use OAuth v2.0 to authenticate to the Quire REST
    API. 


    Quire recommends you to create a dedicated API user with API write access on
    a tenant when authenticating via OAuth, and then create an OAuth client for
    this user. See <a
    href="https://knowledgecenter.zuora.com/CF_Users_and_Administrators/A_Administrator_Settings/Manage_Users/Create_an_API_User"
...


...


...

...

...

...

...
host: quire.io
basePath: /api
...
schemes:
...
paths:
  '/comments/{oid}':
    get:
      tags:
        - comments
      summary: Get a comment
      description: returns the complete comment record.
      operationId: getComment
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of comment that needs to be fetched
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Comment'
    put:
      tags:
        - comments
      summary: Update a comment
      description: >-
        A specific, existing comment can be updated by making a PUT request on
        the URL for that comment.

        Returns the complete updated comment record.
      operationId: updateComment
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: OID of comment that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Comment to update
          required: true
          schema:
            $ref: '#/definitions/UpdateCommentBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Comment'
    delete:
      tags:
        - comments
      summary: Delete a comment
      description: >-
        A specific, existing comment can be deleted by making a DELETE request
        on the URL for that comment.
      operationId: deleteComment
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of comment that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: ok
          examples:
            application/json:
              Success: 'true'
  /organizations:
    get:
      tags:
        - organizations
      summary: Get all organizations
      description: Returns the compact organization records for all organizations
      operationId: getOrganizations
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/OrganizationWithProjects'
            example:
              - id: Develop_group
                oid: Dyh2YkFcu9uLgLFIeN1kB4Ld
                name: Develop group
                initials: DG
                image: 'https://quire.s3.amazonaws.com/oid1/image.jpg'
                iconColor: '#a9cb77'
                createdAt: 2018-12-22T02:06:58.158Z
                createdBy: 6QMKkEPBVWETLWrXqws94ALU
                projects:
                  - id: Marketing_Project
                    oid: GAUpZARpeOjlHqEux6IdUZI1
                    name: Marketing Project
                    initials: MP
                    image: 'https://quire.s3.amazonaws.com/oid3/image.jpg'
                    iconColor: '#a8bf58'
                    createdAt: 2018-12-28T02:06:58.158Z
                    createdBy: 6QMKkEPBVWETLWrXqws94ALU
              - id: Testing_group
                oid: meEXBLRaHFiBTbQd1mJqgPIG
                name: Testing group
                initials: TG
                image: 'https://quire.s3.amazonaws.com/oid2/image.jpg'
                iconColor: '#b3cb60'
                createdAt: 2018-12-27T02:06:58.158Z
                createdBy: 6QMKkEPBVWETLWrXqws94ALU
                projects: []
    post:
      tags:
        - organizations
      summary: Create an organization
      description: Creates a new organization.
      operationId: createOrganization
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Organization to create
          required: true
          schema:
            $ref: '#/definitions/CreateOrganizationBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Organization'
          examples:
            application/json:
              - id: New_Organization
                oid: Dyh2YkFcu9uLgLFIeN1kB4Ld
                name: New Organization
                initials: 'NO'
                image: null
                iconColor: '#a9cb77'
                createdAt: 2018-12-27T12:06:58.158Z
                createdBy: 6QMKkEPBVWETLWrXqws94ALU
  '/organizations/{oid}':
    get:
      tags:
        - organizations
      summary: Get an organization
      description: returns the complete organization record.
      operationId: getOrganization
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of organization that needs to be fetched
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Organization'
    put:
      tags:
        - organizations
      summary: Update an organization
      description: >-
        A specific, existing organization can be updated by making a PUT request
        on the URL for that organization.

        Returns the complete updated organization record.
      operationId: updateOrganization
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of organization that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Organization to update
          required: true
          schema:
            $ref: '#/definitions/UpdateOrganizationBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Organization'
          examples:
            application/json:
              - id: myId
                oid: Dyh2YkFcu9uLgLFIeN1kB4Ld
                name: New name of organization
                initials: 'NO'
                image: null
                iconColor: '#a9cb77'
                createdAt: 2018-12-27T12:06:58.158Z
                createdBy: 6QMKkEPBVWETLWrXqws94ALU
    delete:
      tags:
        - organizations
      summary: Delete an organization
      description: >-
        A specific, existing organization can be deleted by making a DELETE
        request on the URL for that organization.
      operationId: deleteOrganization
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of organization that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: ok
          examples:
            application/json:
              Success: 'true'
  /projects:
    post:
      tags:
        - projects
      summary: Create a project
      description: Creates a new project in an organization.
      operationId: createProject
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Project to create
          required: true
          schema:
            $ref: '#/definitions/CreateProjectBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Project'
  '/projects/{oid}':
    get:
      tags:
        - projects
      summary: Get a project
      description: returns the complete project record for a single project.
      operationId: getProject
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of project that needs to be fetched
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Project'
    put:
      tags:
        - projects
      summary: Update a project
      description: >-
        A specific, existing project can be updated by making a PUT request on
        the URL for that project.

        Returns the complete updated project record.
      operationId: updateProject
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of project that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Project to update
          required: true
          schema:
            $ref: '#/definitions/UpdateProjectBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Project'
    delete:
      tags:
        - projects
      summary: Delete a project
      description: >-
        A specific, existing project can be deleted by making a DELETE request
        on the URL for that project.
      operationId: deleteProject
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of project that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: ok
          examples:
            application/json:
              Success: 'true'
  '/projects/{oid}/comments':
    get:
      tags:
        - projects
      summary: Get project comments
      description: Returns the complete comment record for a single project.
      operationId: getProjectComments
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of project that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Comment'
    post:
      tags:
        - projects
      summary: Add a project comment
      description: Add a new comment in a project.
      operationId: createProjectComment
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of project that comment to
          required: true
          type: string
        - in: body
          name: body
          description: Comment to create
          required: true
          schema:
            $ref: '#/definitions/CreateCommentBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Comment'
  '/projects/{oid}/tags':
    post:
      tags:
        - projects
      summary: Add a project tag
      description: Add a new tag in a project.
      operationId: createProjectTag
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of project
          required: true
          type: string
        - in: body
          name: body
          description: Tag to create
          required: true
          schema:
            $ref: '#/definitions/CreateTagBody'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Tag'
  '/projects/{oid}/tasks':
    get:
      tags:
        - projects
      summary: Get project tasks
      description: Returns the compact task records for all tasks within the given project
      operationId: getTasks
      produces:
        - application/json
      parameters:
        - name: oid
          in: path
          description: Oid of the project in which to search for tasks
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/TaskWithChildren'
...
securityDefinitions:
  api_key:
    type: apiKey
    name: access_token
    in: header
  quire_auth:
    type: oauth2
    authorizationUrl: 'https://quire.io/oauth/authorize'
    flow: accessCode
    scopes:
      'write:projects': modify projects in your account
      'read:projects': read your projects
      basic: to read a user's profile info and media (granted by default)
      comments: to post and delete comments on a user's behalf
definitions:
  Comment:
    type: object
    properties:
      description:
        type: string
        example: Adjust style
      oid:
        type: string
        example: tvax2wFqqiUDqiVF7YSUlQ8n
      pinned:
        type: boolean
        example: false
      createdBy:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
        position: 99
      createdAt:
        type: string
        example: '2018-12-22T02:06:58.158Z'
        position: 99
  CreateCommentBody:
    type: object
    required:
      - description
    properties:
      description:
        type: string
        example: Adjust style
  CreateOrganizationBody:
    type: object
    properties:
      name:
        type: string
        example: New Organization
  CreateProjectBody:
    type: object
    required:
      - organization
    properties:
      name:
        type: string
        example: New Project
      organization:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
        description: The organization oid
  CreateTagBody:
    type: object
    required:
      - name
    properties:
      name:
        type: string
        example: Later
...
  Organization:
    type: object
    properties:
      name:
        type: string
        example: Develop group
      id:
        type: string
        example: Develop_group
      oid:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
      initials:
        type: string
        example: DG
      image:
        type: string
        example: 'https://quire.s3.amazonaws.com/oid/image.jpg'
      iconColor:
        type: string
        example: '#a9cb77'
      createdBy:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
        position: 99
      createdAt:
        type: string
        example: '2018-12-22T02:06:58.158Z'
        position: 99
  OrganizationWithProjects:
    type: object
    properties:
      projects:
        type: array
        items:
          $ref: '#/definitions/Project'
      name:
        type: string
        example: Develop group
      id:
        type: string
        example: Develop_group
      oid:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
      initials:
        type: string
        example: DG
      image:
        type: string
        example: 'https://quire.s3.amazonaws.com/oid/image.jpg'
      iconColor:
        type: string
        example: '#a9cb77'
      createdBy:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
        position: 99
      createdAt:
        type: string
        example: '2018-12-22T02:06:58.158Z'
        position: 99
  Project:
    type: object
    properties:
      name:
        type: string
        example: Marketing Project
      id:
        type: string
        example: Marketing_Project
      oid:
        type: string
        example: GAUpZARpeOjlHqEux6IdUZI1
      initials:
        type: string
        example: MP
      image:
        type: string
        example: 'https://quire.s3.amazonaws.com/oid/image.jpg'
      iconColor:
        type: string
        example: '#a9cb77'
      createdBy:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
        position: 99
      createdAt:
        type: string
        example: '2018-12-22T02:06:58.158Z'
        position: 99
  Tag:
    type: object
    properties:
      name:
        type: string
        example: Later
      oid:
        type: string
        example: PxXtvzXGaC0wzm7uIE92g1j1
      color:
        type: string
        example: '#a9cb77'
      createdBy:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
        position: 99
      createdAt:
        type: string
        example: '2018-12-22T02:06:58.158Z'
        position: 99
...
  TaskWithChildren:
    type: object
    properties:
      oid:
        type: string
        example: 7DQI5OVhLyg2eWaQ7oNvClst
      id:
        type: integer
        format: int32
        example: 12
        position: 1
      name:
        type: string
        example: Design new logo
        position: 2
      state:
        type: integer
        format: int32
        example: 0
        position: 2
      description:
        type: string
        example: Detail about this task
        position: 2
      tags:
        type: array
        position: 2
        items:
          type: string
        example:
          - X6nmx9XjEO2wKbqeB1pRT43C
          - mPAQrYU1qt8wAYAInKRlTnvl
      assignees:
        type: array
        position: 2
        items:
          type: string
        example:
          - 6QMKkEPBVWETLWrXqws94ALU
          - Job4NSW9xK6Owcke8iKj7zyH
      start:
        type: string
        example: '2018-12-20T02:06:58.158Z'
        position: 2
      due:
        type: string
        example: '2018-12-22T02:06:58.158Z'
        position: 2
      tasks:
        type: array
        position: 8
        items:
          $ref: '#/definitions/TaskWithChildren'
        example:
          - oid: co7XeV3ZmvPUpGuS4avHfd7e
            id: '19,'
            name: Draw new logo
            state: 10
            description: Detail about this task
            tags:
              - X6nmx9XjEO2wKbqeB1pRT43C
            assignees:
              - 6QMKkEPBVWETLWrXqws94ALU
            start: null
            due: 2018-12-22T02:06:58.158Z
            tasks: []
            createdAt: 2018-12-24T02:06:58.158Z
            createdBy: Dyh2YkFcu9uLgLFIeN1kB4Ld
      createdBy:
        type: string
        example: Dyh2YkFcu9uLgLFIeN1kB4Ld
        position: 99
      createdAt:
        type: string
        example: '2018-12-22T02:06:58.158Z'
        position: 99
  UpdateCommentBody:
    type: object
    required:
      - description
    properties:
      description:
        type: string
        example: Adjust style
      pinned:
        type: boolean
        example: false
  UpdateOrganizationBody:
    type: object
    properties:
      name:
        type: string
        example: New name of organization
      id:
        type: string
        example: myId
  UpdateProjectBody:
    type: object
    properties:
      name:
        type: string
        example: New name of project
      id:
        type: string
        example: myId
...