# src/main/java/io/quire/api/model/tag/CreateTagBody.java
package io.quire.api.model.tag;

import io.swagger.annotations.ApiModelProperty;

public class CreateTagBody {
    @ApiModelProperty(example = "Later", required = true)
    public String getName() { return null; }
}