# src/main/java/io/quire/api/model/tag/Tag.java
package io.quire.api.model.tag;

import io.quire.api.model.TaggingEntity;
import io.swagger.annotations.ApiModelProperty;

public class Tag extends TaggingEntity {
    @ApiModelProperty(example = "PxXtvzXGaC0wzm7uIE92g1j1")
    public String getOid() {
        return null;
    }

    @ApiModelProperty(example = "Later")
    public String getName() {
        return null;
    }

}